Major refactoring step for glTF 2.0
- Added model classes for glTF 1.0 and 2.0
- Implementation based on glTF 2.0 still pending

+++ jgltf-viewer-lwjgl/src/main/java/de/javagl/jgltf/viewer/lwjgl/GlContextLwjgl.java
@@ -120,24 +120,24 @@ class GlContextLwjgl implements GlContext
         Logger.getLogger(GlContextLwjgl.class.getName());
 
     /**
-     * A buffer that will be used temporarily for the values of 
+     * A buffer that will be used temporarily for the values of
      * integer uniforms. This is a direct buffer that is created
-     * and resized as necessary in {@link #putIntBuffer(int[])} 
+     * and resized as necessary in {@link #putIntBuffer(int[])}
      */
     private IntBuffer uniformIntBuffer = null;
 
     /**
-     * A buffer that will be used temporarily for the values of 
+     * A buffer that will be used temporarily for the values of
      * float uniforms. This is a direct buffer that is created
-     * and resized as necessary in {@link #putFloatBuffer(float[])} 
+     * and resized as necessary in {@link #putFloatBuffer(float[])}
      */
     private FloatBuffer uniformFloatBuffer = null;
 
     /**
      * Put the given values into a direct IntBuffer and return it.
      * The returned buffer may always be a slice of the same instance.
-     * This method is supposed to be called only from the OpenGL thread. 
-     * 
+     * This method is supposed to be called only from the OpenGL thread.
+     *
      * @param value The value
      * @return The IntBuffer
      */
@@ -157,12 +157,12 @@ class GlContextLwjgl implements GlContext
         uniformIntBuffer.flip();
         return uniformIntBuffer;
     }
-    
+
     /**
      * Put the given values into a direct IntBuffer and return it.
      * The returned buffer may always be a slice of the same instance.
-     * This method is supposed to be called only from the OpenGL thread. 
-     * 
+     * This method is supposed to be called only from the OpenGL thread.
+     *
      * @param value The value
      * @return The IntBuffer
      */
@@ -172,7 +172,7 @@ class GlContextLwjgl implements GlContext
         if (uniformFloatBuffer == null || uniformFloatBuffer.capacity() < total)
         {
             uniformFloatBuffer = ByteBuffer
-                .allocateDirect(total * Integer.BYTES)
+                .allocateDirect(total * Float.BYTES)
                 .order(ByteOrder.nativeOrder())
                 .asFloatBuffer();
         }
@@ -182,14 +182,14 @@ class GlContextLwjgl implements GlContext
         uniformFloatBuffer.flip();
         return uniformFloatBuffer;
     }
-    
+
     @Override
     public Integer createGlProgram(
         String vertexShaderSource, String fragmentShaderSource)
     {
         logger.fine("Creating vertex shader...");
-        
-        Integer glVertexShader = 
+
+        Integer glVertexShader =
             createGlShader(GL_VERTEX_SHADER, vertexShaderSource);
         if (glVertexShader == null)
         {
@@ -198,9 +198,9 @@ class GlContextLwjgl implements GlContext
         }
 
         logger.fine("Creating vertex shader DONE");
-        
+
         logger.fine("Creating fragment shader...");
-        Integer glFragmentShader = 
+        Integer glFragmentShader =
             createGlShader(GL_FRAGMENT_SHADER, fragmentShaderSource);
         if (glFragmentShader == null)
         {
@@ -208,7 +208,7 @@ class GlContextLwjgl implements GlContext
             return null;
         }
         logger.fine("Creating fragment shader DONE");
-        
+
         int glProgram  = glCreateProgram();
 
         glAttachShader(glProgram, glVertexShader);
@@ -216,10 +216,10 @@ class GlContextLwjgl implements GlContext
 
         glAttachShader(glProgram, glFragmentShader);
         glDeleteShader(glFragmentShader);
-        
+
         glLinkProgram(glProgram);
         glValidateProgram(glProgram);
-        
+
         int validateStatus = glGetProgram(glProgram, GL_VALIDATE_STATUS);
         if (validateStatus != GL_TRUE)
         {
@@ -228,12 +228,12 @@ class GlContextLwjgl implements GlContext
         }
         return glProgram;
     }
-    
+
     /**
      * Creates an OpenGL shader with the given type, from the given source
-     * code, and returns the GL shader object. If the shader cannot be 
+     * code, and returns the GL shader object. If the shader cannot be
      * compiled, then <code>null</code> will be returned.
-     * 
+     *
      * @param shaderType The shader type
      * @param shaderSource The shader source code
      * @return The GL shader
@@ -242,7 +242,7 @@ class GlContextLwjgl implements GlContext
     {
         int glShader = glCreateShader(shaderType);
         glShaderSource(glShader, shaderSource);
-        glCompileShader(glShader);     
+        glCompileShader(glShader);
         int compileStatus = glGetShader(glShader, GL_COMPILE_STATUS);
         if (compileStatus != GL_TRUE)
         {
@@ -250,19 +250,19 @@ class GlContextLwjgl implements GlContext
         }
         return glShader;
     }
-    
+
     @Override
     public void useGlProgram(int glProgram)
     {
         glUseProgram(glProgram);
     }
-    
+
     @Override
     public void deleteGlProgram(int glProgram)
     {
         glDeleteProgram(glProgram);
     }
-    
+
     @Override
     public void enable(Iterable<? extends Number> states)
     {
@@ -292,7 +292,7 @@ class GlContextLwjgl implements GlContext
             }
         }
     }
-    
+
     @Override
     public int getUniformLocation(int glProgram, String uniformName)
     {
@@ -306,7 +306,7 @@ class GlContextLwjgl implements GlContext
         glUseProgram(glProgram);
         return glGetAttribLocation(glProgram, attributeName);
     }
-    
+
     @Override
     public void setUniformiv(int type, int location, int count, int value[])
     {
@@ -336,18 +336,18 @@ class GlContextLwjgl implements GlContext
                 glUniform3(location, b);
                 break;
             }
-            case GltfConstants.GL_INT_VEC4:   
+            case GltfConstants.GL_INT_VEC4:
             {
                 IntBuffer b = putIntBuffer(value);
                 glUniform4(location, b);
                 break;
             }
             default:
-                logger.warning("Invalid uniform type: " + 
+                logger.warning("Invalid uniform type: " +
                     GltfConstants.stringFor(type));
         }
     }
-    
+
     @Override
     public void setUniformfv(int type, int location, int count, float value[])
     {
@@ -376,19 +376,19 @@ class GlContextLwjgl implements GlContext
                 glUniform3(location, b);
                 break;
             }
-            case GltfConstants.GL_FLOAT_VEC4:   
+            case GltfConstants.GL_FLOAT_VEC4:
             {
                 FloatBuffer b = putFloatBuffer(value);
                 glUniform4(location, b);
                 break;
             }
             default:
-                logger.warning("Invalid uniform type: " + 
+                logger.warning("Invalid uniform type: " +
                     GltfConstants.stringFor(type));
         }
-        
+
     }
-    
+
     @Override
     public void setUniformMatrixfv(
         int type, int location, int count, float value[])
@@ -419,12 +419,12 @@ class GlContextLwjgl implements GlContext
                 break;
             }
             default:
-                logger.warning("Invalid uniform type: " + 
+                logger.warning("Invalid uniform type: " +
                     GltfConstants.stringFor(type));
         }
     }
-    
-    
+
+
     @Override
     public void setUniformSampler(int location, int textureIndex, int glTexture)
     {
@@ -433,14 +433,14 @@ class GlContextLwjgl implements GlContext
         glUniform1i(location, textureIndex);
     }
 
-    
+
     @Override
     public int createGlVertexArray()
     {
         int glVertexArray = glGenVertexArrays();
         return glVertexArray;
     }
-    
+
     @Override
     public void deleteGlVertexArray(int glVertexArray)
     {
@@ -458,10 +458,10 @@ class GlContextLwjgl implements GlContext
         glBufferData(target, bufferViewData, GL_STATIC_DRAW);
         return glBufferView;
     }
-    
+
     @Override
-    public void createVertexAttribute(int glVertexArray, 
-        int target, int glBufferView, int attributeLocation, 
+    public void createVertexAttribute(int glVertexArray,
+        int target, int glBufferView, int attributeLocation,
         int size, int type, int stride, int offset)
     {
         glBindVertexArray(glVertexArray);
@@ -470,30 +470,30 @@ class GlContextLwjgl implements GlContext
             attributeLocation, size, type, false, stride, offset);
         glEnableVertexAttribArray(attributeLocation);
     }
-    
+
     @Override
     public void deleteGlBufferView(int glBufferView)
     {
         glDeleteBuffers(glBufferView);
     }
-    
+
     @Override
     public int createGlTexture(
-        ByteBuffer pixelData, int internalFormat, 
+        ByteBuffer pixelData, int internalFormat,
         int width, int height, int format, int type)
     {
         int glTexture = glGenTextures();
 
         glBindTexture(GL_TEXTURE_2D, glTexture);
         glTexImage2D(
-            GL_TEXTURE_2D, 0, internalFormat, width, height, 
+            GL_TEXTURE_2D, 0, internalFormat, width, height,
             0, format, type, pixelData);
-        
+
         return glTexture;
     }
-    
+
     @Override
-    public void setGlTextureParameters(int glTexture, 
+    public void setGlTextureParameters(int glTexture,
         int minFilter, int magFilter, int wrapS, int wrapT)
     {
         glBindTexture(GL_TEXTURE_2D, glTexture);
@@ -501,33 +501,33 @@ class GlContextLwjgl implements GlContext
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAX_LEVEL, 0);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, minFilter);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, magFilter);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, wrapS); 
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, wrapT); 
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, wrapS);
+        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, wrapT);
     }
-    
+
     @Override
     public void deleteGlTexture(int glTexture)
     {
         glDeleteTextures(glTexture);
     }
-    
+
     @Override
     public void renderIndexed(
-        int glVertexArray, int mode, int glIndicesBuffer, 
+        int glVertexArray, int mode, int glIndicesBuffer,
         int numIndices, int indicesType, int offset)
     {
         glBindVertexArray(glVertexArray);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, glIndicesBuffer);
         glDrawElements(mode, numIndices, indicesType, offset);
     }
-    
+
     @Override
     public void renderNonIndexed(int glVertexArray, int mode, int numVertices)
     {
         glBindVertexArray(glVertexArray);
         glDrawArrays(mode, 0, numVertices);
     }
-    
+
     @Override
     public void setBlendColor(float r, float g, float b, float a)
     {
@@ -603,16 +603,16 @@ class GlContextLwjgl implements GlContext
 
     /**
      * For debugging: Print shader log info
-     * 
+     *
      * @param id shader ID
      */
-    private void printShaderLogInfo(int id) 
+    private void printShaderLogInfo(int id)
     {
         IntBuffer infoLogLength = ByteBuffer.allocateDirect(4)
             .order(ByteOrder.nativeOrder())
             .asIntBuffer();
         glGetShader(id, GL_INFO_LOG_LENGTH, infoLogLength);
-        if (infoLogLength.get(0) > 0) 
+        if (infoLogLength.get(0) > 0)
         {
             infoLogLength.put(0, infoLogLength.get(0)-1);
         }
@@ -627,20 +627,20 @@ class GlContextLwjgl implements GlContext
         {
             logger.warning("shader log:\n"+infoLogString);
         }
-    }    
+    }
 
     /**
      * For debugging: Print program log info
-     * 
+     *
      * @param id program ID
      */
-    private void printProgramLogInfo(int id) 
+    private void printProgramLogInfo(int id)
     {
         IntBuffer infoLogLength = ByteBuffer.allocateDirect(4)
             .order(ByteOrder.nativeOrder())
             .asIntBuffer();
         glGetProgram(id, GL_INFO_LOG_LENGTH, infoLogLength);
-        if (infoLogLength.get(0) > 0) 
+        if (infoLogLength.get(0) > 0)
         {
             infoLogLength.put(0, infoLogLength.get(0)-1);
         }
@@ -649,13 +649,13 @@ class GlContextLwjgl implements GlContext
             .order(ByteOrder.nativeOrder());
         glGetProgramInfoLog(id, infoLogLength, infoLog);
 
-        String infoLogString = 
+        String infoLogString =
             Charset.forName("US-ASCII").decode(infoLog).toString();
         if (infoLogString.trim().length() > 0)
         {
             logger.warning("program log:\n"+infoLogString);
         }
-    }    
+    }
 
 
 }
